requirements.txt

reportlab==4.0.9
gunicorn==21.2.0
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10